+++ src/sentry/integrations/jira/views/__init__.py
@@ -2,8 +2,8 @@ UNABLE_TO_VERIFY_INSTALLATION = "Unable to verify installation"
 
 from .base import JiraSentryUIBaseView
 from .extension_configuration import JiraExtensionConfigurationView
-from .issue_hook import JiraSentryIssueDetailsView
-from .ui_hook import JiraSentryInstallationView
+from .sentry_installation import JiraSentryInstallationView
+from .sentry_issue_details import JiraSentryIssueDetailsView
 
 __all__ = (
     "JiraSentryUIBaseView",
