+++ src/sentry/integrations/jira/views/sentry_installation.py
