+++ src/sentry/integrations/jira/views/sentry_issue_details.py
